+++ web/source/components/home.jsx
@@ -1,3 +1,5 @@
+var ReactTooltip = require('../../../node_modules/react-tooltip/standalone/react-tooltip.min.js');
+
 var Component = React.createClass({
 	render: function () {
 		return (
@@ -5,7 +7,7 @@ var Component = React.createClass({
 				<div className="jumbotron custom-background custom-background-home">
 					<div className="container">
 						<h1>Manta Config Engine App</h1>
-						<p>Graphical Userinterface to generate autoexec files for Dota 2.</p>
+						<p>A web-application to generate autoexec files for Dota 2.</p>
 						<a className="btn btn-success" href="#/editor">Start Editing</a>
 					</div>
 				</div>
@@ -18,10 +20,28 @@ var Component = React.createClass({
 									<li>Click on the keys you want to customize</li>
 									<li>Consider customizing additional layouts</li>
 									<li>Click download</li>
-									<li>Copy the contents of the zip-Archive to your autoexec folder</li>
-									<li>Restart Dota 2 or enter "exec autoexec.cfg" into your console</li>
+									<li>
+										Copy the contents of the zip-Archive to your autoexec folder <i className="glyphicon glyphicon-info-sign" data-tip data-for="home-info-autoexec"/>
+									<ReactTooltip id="home-info-autoexec" place="top" type="light" effect="solid">
+											See Below
+										</ReactTooltip>
+									</li>
+									<li>Restart Dota 2 or type <code>exec autoexec.cfg</code> into the console</li>
 								</ul>
-								Tip: You need to remove all keybindings for this to work
+								<strong>Tip:</strong> You need to remove conflicting ingame-keybinds <i className="glyphicon glyphicon-info-sign" data-tip data-for="home-info-keybinds"/>
+								<ReactTooltip id="home-info-keybinds" place="top" type="light" effect="solid">
+									<p style={{textAlign: 'center'}}>
+										Manta <strong>can't use keys bound in the Dota 2 settings</strong>.
+										<br/>
+										To unset a key in Dota, just go to the settings,
+										<br/>
+										click the keybind you need to remove and hit <kbd>PrintScr</kbd>
+										<br/>
+										If you don't have <kbd>PrintScr</kbd> just press <kbd>Any Key</kbd>.
+										<br/>
+										Repeat this for every key you need to unbind.
+									</p>
+								</ReactTooltip>
 							</p>
 							<h2>Default Autoexec Folders</h2>
 							<div className="media">
@@ -38,7 +58,7 @@ var Component = React.createClass({
 									<img src="https://openclipart.org/download/159331/MoonBook.svg" width="64"/>
 								</div>
 								<div className="media-body">
-									<h4 className="media-heading">Mac OSX</h4>
+									<h4 className="media-heading">Mac OS X</h4>
 									/Users/YourUserName/Library/Application Support/Steam/steamapps/common/dota 2 beta/game/dota/cfg
 								</div>
 							</div>
@@ -55,18 +75,25 @@ var Component = React.createClass({
 						<div className="col-sm-4">
 							<h2>What Is This?</h2>
 							<p>
-								Manta is a tool, which can generate <b>advanced configuration files</b> for <a href="http://dota2.com">Dota 2</a>.
-								These can for example allow you to use QuickCast, while being able to self-target spells &amp; items without using your mouse.
-								Effectively <b>combining the advantages</b> of both Quick- and NormalCast.
+								Manta is a tool, which can generate <a href="http://wiki.teamliquid.net/dota2/Console_Commands">advanced configuration files</a> for <a href="http://dota2.com">Dota 2</a>.
 							</p>
-							<h2>Things Manta Can Do</h2>
+							<div className="media">
+								<div className="media-left media-middle">
+									<i className="glyphicon glyphicon-flash" style={{fontSize: '4em'}}/>
+								</div>
+								<div className="media-body">
+									<h4 className="media-heading">Awesome QuickCast</h4>
+									Combine the advantages of Quick- and NormalCast. Use <kbd><kbd>Space</kbd>+<kbd>Q</kbd></kbd> to cast <kbd>Q</kbd> on yourself.
+								</div>
+							</div>
 							<div className="media">
 								<div className="media-left media-middle">
 									<i className="glyphicon glyphicon-duplicate" style={{fontSize: '4em'}}/>
 								</div>
 								<div className="media-body">
 									<h4 className="media-heading">Multiple Layouts</h4>
-									The default <a href="#/editor">layout</a> of Manta uses this to allow you to press <kbd>Space</kbd>+<kbd>Q</kbd> to self-cast your Q.
+									You can create as many <a href="#/editor">layouts</a> as you want. That allows you to use nearly unlimited keyboard shortcuts.
+									Call <strong><i className="glyphicon glyphicon-play"/> Missing Top</strong> with <kbd><kbd>Space</kbd>+<kbd>1</kbd></kbd> and much more.
 								</div>
 							</div>
 							<div className="media">
@@ -75,7 +102,7 @@ var Component = React.createClass({
 								</div>
 								<div className="media-body">
 									<h4 className="media-heading">Multiple Chatwheels</h4>
-									The default layout of Manta has two Chatwheels. One bound to <kbd>Y</kbd> and one bound to <kbd>Space</kbd>+<kbd>Y</kbd>.
+									The default layout of Manta has two Chatwheels. One bound to <kbd>Y</kbd> and one bound to <kbd><kbd>Space</kbd>+<kbd>Y</kbd></kbd>.
 									You can easily customize them in the <a href="#/chatwheels">Chatwheel Manager</a>
 								</div>
 							</div>
@@ -85,9 +112,8 @@ var Component = React.createClass({
 								</div>
 								<div className="media-body">
 									<h4 className="media-heading">Advanced Settings</h4>
-									The <a href="#/settings">Settings</a> section of Manta allows you to customize nearly everything.
-									It includes settings which are not available in the ingame settings.
-									You can even enable hidden Dota features like "Force Movement Direction" which allows you to turn your hero without moving when pressing <kbd>Alt</kbd>.
+									The <a href="#/settings">Settings</a> section of Manta allows you to customize nearly everything,
+									including hidden settings and features like <strong>Force Movement Direction</strong> which allows you to turn your hero without moving when pressing <kbd>Alt</kbd>.
 								</div>
 							</div>
 						</div>
@@ -116,6 +142,8 @@ var Component = React.createClass({
 										<a href="https://github.com/dodekeract/manta-config-engine/tree/master/documentation/README.md">Documentation</a>
 										<br/>
 										<a href="https://github.com/dodekeract/manta-config-engine">GitHub Repository</a>
+										<br/>
+										<a href="https://npmjs.org/package/dota2-manta-config-engine">View On NPM</a>
 									</div>
 								</div>
 								<div className="media">
@@ -127,6 +155,21 @@ var Component = React.createClass({
 										<a href="https://github.com/dodekeract/manta-config-engine/tree/master/documentation/README.md">Documentation</a>
 										<br/>
 										<a href="https://github.com/dodekeract/manta-config-engine">GitHub Repository</a>
+										<br/>
+										<a href="https://npmjs.org/package/dota2-manta-config-engine">View On NPM</a>
+									</div>
+								</div>
+								<div className="media">
+									<div className="media-left media-middle">
+										<i className="glyphicon glyphicon-link" style={{fontSize: '4em'}}/>
+									</div>
+									<div className="media-body">
+										<h4 className="media-heading">Additional Links</h4>
+										<a href="https://gitter.im/dodekeract/manta-config-engine">Gitter Chatroom</a>
+										<br/>
+										<a href="https://github.com/dodekeract/manta-config-engine-app/blob/master/documentation/CHANGELOG.md">Changelog</a>
+										<br/>
+										<a href="https://github.com/dodekeract/manta-config-engine-app/blob/master/documentation/LICENSE.md">License</a>
 									</div>
 								</div>
 							</p>