+++ documentation/CHANGELOG.md
@@ -1,8 +1,23 @@
 # Changelog
 
+## 1.3.4
+- home
+	- added tooltips
+	- added additional links
+- layout-editor
+	- fixed tooltips sometimes being duplicated
+	- use sharper edges on keys
+	- changed legend layout & style
+- preset-viewer
+	- added reset button
+- fixed tabs losing focus
+- moved / to /home
+- code-cleanup
+
 ## 1.3.3
 - improved home-section
 - added icons to tab bar
+- fixed a memory leak in the new bind-changer
 
 ## 1.3.2
 - fixed a critical bug regarding git auto-tagging
@@ -12,7 +27,7 @@
 
 ## 1.3.0
 - completely re-worked keybind-changer
-- fixed tooltips sometimes not showing up in the layout editor
+- fixed tooltips sometimes not showing up in the layout-editor
 
 ## 1.2.2
 - enabled automatic git tagging of npm versions